Extract time-series at several punctual locations
Have you ever tried to extract long time-series at high temporal resolution from Google Earth Engine?

If you have, you may have found that downloading everything in a csv file into your Google Drive takes a while, leading sometimes to a crash. You can print out directly to the console the information about an Image or Feature Collection, but the console printout is limited to 5000 elements. 
In addition, you might prefer manipulating time-series as dataframes or dictionaries in Python, or R. The Earth Engine Python APi is available as a Python package than can be installed locally. The package is called "ee". It must be imported and initialized for each new Python session and script. This package allows you to work in Spyder (for example) with a hybrid syntax of Javascript and Python. After having followed the necessary few steps to connect your google account to your machine, you can easily load into Spyder the GEE products using that hybrid syntax.
Yet, there is no function allowing to transform the typical GEE formats such as Image Collections or Feature Collections into Python handy formats such as dictionnaries or pandas dataframes. In order to achieve this, from what I know, there is no getting around using the getInfo() GEE function. 
Because the client doesn't know what's in server-side objects, JavaScript functionality such as conditionals and for-loops does not work with them. Sam with Python conditionals and for-loops. For that reason, and to avoid synchronous calls to getInfo(), Google recommends to use server functions to the extent possible. Finally, getInfo() is also restricted to 5000 elements.

This function (and its auxiliary functions) I created allows one to download easily from Google Earth Engine directly into a dictionnary of yearly dataframes time-series at multiple points in space. The function is coded so as to download in chuncks of 4500 elements, nevermind the time-resolution, the number of bands or the number of points to download the products at (up to a certain limit. You can imagine why it would not work if you already have > 4500 points to download information from. This latter restrictions can also be worked around of course in further code improvements).

The getInfo() function is iteratively used, and requires bandwidth. Therefore, the computation speed is (amongst other things) a function of your internet speed. The function may also crash if at some point your internet connection is turned off. To avoid losing everything in these cases, I coded the function so that the time-series are saved in yearly chunks into a dictionary locally saved at the directory you specify. If the function crashes, you can just run it again with the missing years. 

I coded this function to download rainfall time series at several locations. Any improvement or comments are very welcome.
Hope this helps.